recvAllSynapticInput uses observer-pattern messages
recvAllSynapticInput sends a DeliverInputMessage
instead of directly calling HyPerConn::deliver().

During the communicateInitInfo stage, connections
send a LayerAddInputMessage to their post layers,
and when recvSynapticInput is called, the layer
sends a DeliverInputMessage to each object that
had added itself.

The HyPerLayer code no longer needs any class in
the connection hierarchy (HyPerConn still depends
on HyPerLayer to check that preLayerName and
postLayerName refer to HyPerLayers, and to find
the GSyn and datastore buffers).